ENSITECH-285 add barrel file for component
